JSONified cases
This is the beginning of getting questions to be generated.  I'm
planning to also JSONify the aspects for the first release, but other
categories will need to wait a little.

At some point, i should also explain the format for these files, because
it's definitely got some problems that ought to be solved.  It works for
now, though.