Redid main loop and rendering a bit
It's now a lot more practical, and i like it.  I also removed a couple
instances of worrying too much about cosmic rays.  Neither of these took
very long; what _did_ take too fucking long was me trying and failing to
get scaling to work so that the window could be resizable.  Scaling
wasn't too hard, but getting mouse clicks to map to the right places was
a fucking nightmare, so i eventually just gave up on the whole
ordeal…_after_ sinking several hours and a lot of sanity into it.  If
someone else wants to try it, that would be just fine by me.

Anyway.  The next step is to get JSON parsing working so that we can
(in addition to properly checking the validity of the files) actually
construct the questions that are the basis of this whole thing.
(There's also a particular block of code that ought to be rewritten…but
priorities first!)