Fixed JSON, added timer to menu, buttons now work
The JSON files are now actually well-formed, and the "i"s are now
dotless where appropriate to reflect modern usage.

There's now a timer button on the menu screen.  It'll also be on the
questions screen once i get that implemented.  The timer is communicated
between the two modes by the userdata in the surface.

The buttons now go to the questions mode and load the correct file, but
since the questions mode isn't implemented yet, it just flashes back to
the menu screen.

All of this done before lunchtime.  Horray!  This is a good day.  :-)

Next step is to get the questions screen up and running.  I'm going to
use some hardcoded questions and answers at first so that i don't have
to worry about JSON at the same time.  After that, i'll get JSON parsing
up and working with the questions.  Then, it should be on to release
N.1-beta.1!