It finally works!
This one has done…a _lot_.  But the application finally works now!
Horray!

There are a few small problems that i'll fix before the first release,
and a couple of other bigger problem.  I'll talk about these at the end
of the message.

First:  The JSON files now use escape sequences instead of the control
characters directly, to be compliant with the specification.  There's
also now a JSON library that's being used to parse these—this is
something that i'm very unhappy about.  I particularly don't like how it
parses the file into a completely new data structure to be interacted
with, despite the fact that the JSON file is _literally right the fuck
there_ and we could just parse that directly instead and save so much
space.  I also don't like the data structure to which it parses it—it's
so cumbersome and verbose to do address things.  Because of this bloat,
i've also put a warning in README.md that this is a rather hefty
application (which really should always have been there, considering the
use of SDL).  At least there's one silver lining:  determining which
assets are JSON files has gotten a lot more robust.

Secondly, i went through the codebase and did a bunch of reorganization
of things to reduce repetition, along with some general refactoring of
stuff.  It's still a fucking terrible codebase, but it's at least kinda
sorta slightly better now, i guess.  This was done in part due to
deperation induced by insanity induced by debugging of…

The biggest and most obvious thing:  The questions now load from the
JSON files correctly.  This took a long-ass time to do because i kept
running into extremely fucking bizarre bugs and glitches that were very
difficult to diagnose.  Some of them were seemingly completely random,
some of them were caused by undefined behavior, and some of them would
disappear when i tried to use a debugger to analyse them (i've learned
that this is called a "heisenbug", which is just a bloody wonderful
name).  But i've managed to fix them, and now the program hopefully will
neither crash nor freeze!

There are still a few problems that i think i can fix before i release
version N.1-beta.0:
	- A few issues with text wrapping, such as the response when you
	get a question wrong not wrapping, the questions and answers
	wrapping when they don't need to sometimes, and most obviously
	the fact that wrapped text isn't centered.
	- Very thin answers (such as "ı") can be difficult to select.

And some more problems that will need to wait until after version
N.1-beta.0, and perhaps until after N.1:
	- The timer for questions doesn't work.
	- The window can't be scaled.  I tried fixing this once, but i
	wasn't able to reconcile discrepencies in the apparent and
	actual click locations.
	- Some of the glyphs that Ithkuil uses aren't supported by the
	font that the program is using, Barlow Condensed.
	- Bloat.  The three main sources of this are Barlow Condensed,
	the JSON library, and most egregiously SDL.  This will be a
	difficult and complicated issue to fix.
	- The format for the assets works for now, but it could
	definitely do with some improvements.  (I'm thinking this may
	evolve into a separate project dedicated solely to the
	digitization of the Ithkuil language.)
	- Perhaps most obviously, only the aspects and cases have been
	digitized yet.  Digitizing the whole language will take a good
	long while (most of which will be taken up by the lexicon and
	the VxCs affixes), and determining what the best way to organize
	things is may be difficult.

I'm nervous about this.  I don't know _why_ i'm nervous—after all, i've
known that the first release was coming for a while now.  I guess
there's a difference between imagining doing something and actually
doing it—that doesn't bode too well for my future.  Still, i'm going to
try to overcome this and publicly release this.

Hooooooooooooo boy.